Fix: CI failures from PHPCS warning cleanup (#77)
- Fix type mismatch in match outcome: cast club IDs to int for strict comparison
- Fix widget show_abbr: use boolean check instead of integer strict comparison
- Fix lineup-row: use truthy check for player number meta
- Initialize $b array before foreach in sort functions to fix undefined variable
- Fix league-table/standings: use ! empty() for shortcode attribute checks
- Fix roster meta boxes: use ! empty() instead of null !== for array checks
- Fix table-stats meta box: same ! empty() pattern
- Fix admin-post-types: move phpcs:ignore to end of line to preserve translators comments
- Fix admin.php: sanitize $_SERVER['SCRIPT_FILENAME'] with wp_unslash + sanitize_text_field
- Fix staff-details/player-details: dedent array items after removing commented-out if blocks
- Update PHPStan baseline: WPCM_VERSION count 2→3, W3TC_FILE count 2→1
- Remove resolved PHPStan baseline entries for $b and $link_club

+++ includes/wpcm-match-functions.php
@@ -108,8 +108,8 @@ add_action( 'wp_ajax_item_sort', 'wpcm_match_players_item_order' );
  */
 function wpcm_get_match_outcome( $post ) {
 
-	$club      = get_default_club();
-	$home_club = get_post_meta( $post, 'wpcm_home_club', true );
+	$club      = (int) get_default_club();
+	$home_club = (int) get_post_meta( $post, 'wpcm_home_club', true );
 	$walkover  = get_post_meta( $post, '_wpcm_walkover', true );
 	$postponed = get_post_meta( $post, '_wpcm_postponed', true );
 	if ( get_option( 'wpcm_sport' ) !== 'cricket' ) {
